feat: Render player after initializing it
Since player can only move in discrete steps, the math is encapsulated.

The game world is a — in this case, 5x6 — grid.
Player moves from grid to grid, never halfway between.

Whenever we want to render player in a specified cell,
we have to convert its cell coordinate to canvas coordinate.
For example, to render player at its initial location:
We have Cell Coordinate = (2, 5), and need to convert it to
Canvas Coordinate = (2 * 101, (5*83) - 30) = (202, 385).

This could be done directly in update(), but given its complexity,
better be abstracted away; setCanvasCoordinate() handles exactly that.