feat: Enemy reincarnates after falling off game world
Game world is now horizontally cyclic for Enemy instances.

Mathematically, we treat it as if there are invisible columns of cells
on the left and right of the visible game world. One column on each side.

When an Enemy instance is about to 'fall off the face of the earth' —
in our case, reaching the center of corresponding invisible cell on the right —
it's reincarnated at the center of corresponding invisible cell on the left.